chore(hooks): Extract add to basket functionality as a hook
